+++ backend/src/baserow/config/settings/base.py
@@ -155,6 +155,12 @@ AUTH_PASSWORD_VALIDATORS = [
     },
 ]
 
+# We need the `AllowAllUsersModelBackend` in order to respond with a proper error
+# message when the user is not active. The only thing it does, is allowing non active
+# users to authenticate, but the user still can't obtain or use a JWT token or database
+# token because the user needs to be active to use that.
+AUTHENTICATION_BACKENDS = ["django.contrib.auth.backends.AllowAllUsersModelBackend"]
+
 
 # Internationalization
 # https://docs.djangoproject.com/en/2.2/topics/i18n/
@@ -276,7 +282,7 @@ EMAIL_BACKEND = "djcelery_email.backends.CeleryEmailBackend"
 
 if os.getenv("EMAIL_SMTP", ""):
     CELERY_EMAIL_BACKEND = "django.core.mail.backends.smtp.EmailBackend"
-    # EMAIL_SMTP_USE_TLS OR EMAIL_SMTP_USE_TLS for backwards compatibilty after
+    # EMAIL_SMTP_USE_TLS OR EMAIL_SMTP_USE_TLS for backwards compatibility after
     # fixing #448.
     EMAIL_USE_TLS = bool(os.getenv("EMAIL_SMTP_USE_TLS", "")) or bool(
         os.getenv("EMAIL_SMPT_USE_TLS", "")